Fix: Cart Product Quantity
- This resolves issue #56
- The cart page will now update the prices according to the quantity amount
- Can now view the shopping cart page if no products are in the shopping cart
- Added a new class to handle the quantity amount in the shopping cart
- You can add an item to the cart wilout reloading the entire page
- When adding a product to your cart, you will no longer be redirected.

+++ MVC_OnlineShop/Models/Product.cs
@@ -8,7 +8,7 @@ namespace MVC_OnlineShop.Models {
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; }
         public string Image_Url { get; set; }
         public ProductType Type { get; set; }
